model_training/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
tqdm>=4.62.0
Pillow>=8.3.0